requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
python-dotenv
mysql-connector-python
python-jose
bcrypt